requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pdfplumber==0.11.4
playwright==1.49.1
apscheduler==3.10.4
pyyaml==6.0.2
httpx==0.28.1
